refactor(): Remove controller
